requirements.txt
streamlit>=1.37
pillow